+++ Model/Plant/Reproductive/GrainPartGN.cpp
@@ -61,7 +61,7 @@ void fruitGrainPartGN::onInit1(protocol::Component *system)
 float fruitGrainPartGN::grainWt(void)
    //===========================================================================
 {
-   return divide (Total.DM(), gGrain_no, 0.0);
+   return (float)divide (Total.DM(), gGrain_no, 0.0);
 }
 
 void fruitGrainPartGN::get_grain_size(protocol::Component *system, protocol::QueryValueData &qd)
@@ -163,10 +163,10 @@ void fruitGrainPartGN::readSpeciesParameters(protocol::Component *system, vector
    {
    fruitGrainPart::readSpeciesParameters(system, sections);
 
-   rel_grainfill.read(scienceAPI, "x_temp_grainfill", "oC", 0.0, 80.0
-                                   , "y_rel_grainfill", "-", 0.0, 1.0);
-   rel_grain_n_fill.read(scienceAPI, "x_temp_grain_n_fill", "oC", 0.0, 40.0
-                                   , "y_rel_grain_n_fill", "-", 0.0, 1.0);
+   rel_grainfill.read(scienceAPI, "x_temp_grainfill", "oC", 0.0f, 80.0f
+                                   , "y_rel_grainfill", "-", 0.0f, 1.0f);
+   rel_grain_n_fill.read(scienceAPI, "x_temp_grain_n_fill", "oC", 0.0f, 40.0f
+                                   , "y_rel_grain_n_fill", "-", 0.0f, 1.0f);
    scienceAPI.read("potential_grain_n_filling_rate", cPotential_grain_n_filling_rate, 0.0f, 1.0f);
    scienceAPI.read("minimum_grain_n_filling_rate", cMinimum_grain_n_filling_rate, 0.0f, 1.0f);
    scienceAPI.read("crit_grainfill_rate", cCrit_grainfill_rate, 0.0f, 1.0f);
@@ -181,7 +181,7 @@ void fruitGrainPartGN::update(void)
    float dlt_grain_no_lost  = gGrain_no * plant->population().DyingFractionPlants();
    gGrain_no -= dlt_grain_no_lost;
    
-   gGrain_no = gGrain_no * (1.0 - GrainKillFraction);
+   gGrain_no = (float)(gGrain_no * (1.0 - GrainKillFraction));
    GrainKillFraction = 0.0;
    Debug("meal.GrainNo=%f", gGrain_no);
